requirements.txt
anthropic>=0.40.0
python-pptx>=1.0.0
markdown>=3.5
weasyprint>=62.0
requests>=2.31.0
